feat(lotto): 당첨 번호 입력받기 + 보너스 번호 입력받기 기능 구현
