+++ src/main/java/lotto/domain/Calculator.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import lotto.Lotto;
-
 public class Calculator {
 
     public void countCorrectLuckyNumbers(){
